fix(Snap): Make Avg for oos

+++ apps/snap/serializers/out_of_stock_serializers.py
@@ -75,9 +75,9 @@ class UpdateOutOfStockSnapSerializer(OutOfStockSnapSerializer):
 
 class OverviewOutOfStockSnapReport(serializers.Serializer):
     sku = serializers.CharField(source='sku_name')
-    available = serializers.FloatField()
-    not_available = serializers.FloatField()
-    less = serializers.FloatField()
+    available = serializers.DecimalField(max_digits=3, decimal_places=1)
+    not_available = serializers.DecimalField(max_digits=3, decimal_places=1)
+    less = serializers.DecimalField(max_digits=3, decimal_places=1)
 
 
 class OutOfStockSnapReportSerializer(serializers.Serializer):
